Removed API-specific default values from gateway class (temporary).
All the default values in the gateway class *must* be supported with
setters and getters in all request messages that the gateway supports;
the core tests insist on this. Most of those parameters are irrelevant
for the complete*, the acceptNotification and the other backend functions,
so make no sense to implemement. Those messages will probably have to be
moved to a "Backend" gateway class so the default parameters match the
supported message parameters.